various build fixes

+++ woosh-photo-app/SecondViewController.m
@@ -147,7 +147,7 @@ int req_type = REQUEST_TYPE_NONE;
             
                 NSTimeInterval interval = [offerEnd timeIntervalSinceNow];
                 NSInteger time = interval;
-                cell.remainingTimeLabel.text = [NSString stringWithFormat:@"%d:%02d remaining for offer", time / 60, time % 60];
+                cell.remainingTimeLabel.text = [NSString stringWithFormat:@"%f:%02d remaining for offer", time / 60.0, (int)time % 60];
                 
             } else {
                 
@@ -190,7 +190,7 @@ int req_type = REQUEST_TYPE_NONE;
                 
                     NSTimeInterval interval = [offerEnd timeIntervalSinceNow];
                     NSInteger time = interval;
-                    cell.remainingTimeLabel.text = [NSString stringWithFormat:@"%d:%d remaining for offer", time / 60, time % 60];
+                    cell.remainingTimeLabel.text = [NSString stringWithFormat:@"%f:%02d remaining for offer", time / 60.0, (int)time % 60];
 
                 } else {
